pyproject.toml
[project]
name = "fm-adem"
version = "0.1.0"
description = "\"adem reports\""
package-mode = false
authors = [
    {name = "Kanat",email = "Kanat@gm.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
